feat(frontend): FE UI cleanup
