+++ web/src/components/ResizablePanels/FillPanel.tsx
@@ -1,5 +1,5 @@
 import * as Spaces from 'react-spaces';
 
-const FillPanel: React.FC = ({children}) => <Spaces.Fill>{children}</Spaces.Fill>;
+const FillPanel: React.FC = ({children}) => <Spaces.Fill style={{overflow: 'scroll'}}>{children}</Spaces.Fill>;
 
 export default FillPanel;
