+++ web/src/components/RunDetailTrace/RunDetailTrace.styled.ts
@@ -2,10 +2,11 @@ import {CloseCircleFilled} from '@ant-design/icons';
 import {Typography} from 'antd';
 import styled from 'styled-components';
 
-export const Container = styled.div`
+export const Container = styled.div<{$isTimeline: boolean}>`
   display: flex;
   height: 100%;
   width: 100%;
+  min-width: ${({$isTimeline}) => $isTimeline && '1000px'};
 `;
 
 export const SearchContainer = styled.div`
