+++ web/src/components/RunDetailTrace/Visualization.tsx
@@ -9,7 +9,7 @@ import TraceSelectors from 'selectors/Trace.selectors';
 import TestRunService from 'services/TestRun.service';
 import Trace from 'models/Trace.model';
 import {TTestRunState} from 'types/TestRun.types';
-import TimelineV2 from 'components/Visualization/components/Timeline/TimelineV2';
+import TimelineV2 from 'components/Visualization/components/Timeline/Timeline';
 import {VisualizationType} from './RunDetailTrace';
 import TraceDAG from './TraceDAG';
