+++ web/src/components/Visualization/components/Timeline/Timeline.tsx
@@ -1,31 +1,36 @@
-import {ParentSize} from '@visx/responsive';
 import {NodeTypesEnum} from 'constants/Visualization.constants';
 import Span from 'models/Span.model';
-import * as S from './Timeline.styled';
-import Visualization from './Visualization';
-import Navigation from '../Navigation';
+import {useRef} from 'react';
+import {FixedSizeList as List} from 'react-window';
+import NavigationWrapper from './NavigationWrapper';
+import TimelineProvider from './Timeline.provider';
+import ListWrapper from './ListWrapper';
 
 export interface IProps {
-  isMatchedMode: boolean;
-  matchedSpans: string[];
   nodeType: NodeTypesEnum;
-  onNavigateToSpan(spanId: string): void;
-  onNodeClick(spanId: string): void;
-  selectedSpan: string;
   spans: Span[];
-  width?: number;
+  onNavigate(spanId: string): void;
+  onClick(spanId: string): void;
+  matchedSpans: string[];
+  selectedSpan: string;
 }
 
-const Timeline = (props: IProps) => {
-  const {isMatchedMode, matchedSpans, onNavigateToSpan, selectedSpan} = props;
+const Timeline = ({nodeType, spans, onClick, onNavigate, matchedSpans, selectedSpan}: IProps) => {
+  const listRef = useRef<List>(null);
 
   return (
-    <S.Container $showMatched={isMatchedMode}>
-      <Navigation matchedSpans={matchedSpans} onNavigateToSpan={onNavigateToSpan} selectedSpan={selectedSpan} />
-      <ParentSize parentSizeStyles={{height: '100%', overflowY: 'scroll', paddingTop: 32, width: '100%'}}>
-        {({width}) => <Visualization {...props} width={width} />}
-      </ParentSize>
-    </S.Container>
+    <TimelineProvider
+      onClick={onClick}
+      onNavigate={onNavigate}
+      matchedSpans={matchedSpans}
+      selectedSpan={selectedSpan}
+      listRef={listRef}
+      nodeType={nodeType}
+      spans={spans}
+    >
+      <NavigationWrapper />
+      <ListWrapper listRef={listRef} />
+    </TimelineProvider>
   );
 };
